Use generator only in createTileMap.

+++ src/Tilemap.hpp
@@ -15,7 +15,7 @@ struct Size {
 };
 
 struct Tile {
-    int id;
+    uint8_t id;
     uint8_t index;
 };
 
@@ -24,7 +24,7 @@ public:
     Tilemap(int width, int height, int tile_width, int tile_height);
 
     void setTexture(std::shared_ptr<Texture2D> texture);
-    void setTile(int x, int y, int id, uint8_t index);
+    void setTile(int x, int y, uint8_t id, uint8_t index);
     Tile getTile(int x, int y);
 
     void drawTile(float x, float y, int index);
